fix: resolve all CI errors - remove unused Color4Extensions usings, fix Colour4/Matrix3x2/Vector2/Key migration issues
Agent-Logs-Url: https://github.com/winnerspiros/osu/sessions/d5e0a29e-7e36-48de-ab95-334865fa54ac

+++ osu.Game/Rulesets/Objects/BezierConverter.cs
@@ -225,7 +225,7 @@ namespace osu.Game.Rulesets.Objects
                 {
                     for (int i = n; i > j; i--)
                     {
-                        arc[i] = arc[i] * tf + arc[i - 1] * (1 - tf);
+                        arc[i] = arc[i] * (float)tf + arc[i - 1] * (float)(1 - tf);
                     }
                 }
